+++ packages/terra-overlay/tests/nightwatch/Overlay/OverlayTests.jsx
@@ -10,8 +10,9 @@ const OverlayTests = () => (
       <li><Link to="/tests/overlay-tests/container">Overlay - Relative to Container</Link></li>
       <li><Link to="/tests/overlay-tests/custom-content">Overlay - Custom Content</Link></li>
       <li><Link to="/tests/overlay-tests/on-request-close">Overlay - On Request Close</Link></li>
+      <li><Link to="/tests/overlay-tests/disable-close-on-esc">Overlay - Disabled Close On Escape</Link></li>
       <li><Link to="/tests/overlay-tests/fullscreen-scrollable">Overlay - Fullscreen Scrollable</Link></li>
-      <li><Link to="/tests/overlay-tests/container-scrollable">Overlay - Relativeto Container Scrollable</Link></li>
+      <li><Link to="/tests/overlay-tests/container-scrollable">Overlay - Relative to Container Scrollable</Link></li>
       <li><Link to="/tests/overlay-tests/light-themed">Overlay - Light Themed</Link></li>
       <li><Link to="/tests/overlay-tests/dark-themed">Overlay - Dark Themed</Link></li>
       <li><Link to="/tests/overlay-tests/clear-themed">Overlay - Clear Themed</Link></li>